Use float64 instead of float32 for metrics of Float type

+++ metrics/core/types.go
@@ -86,7 +86,7 @@ func (self *UnitsType) String() string {
 
 type MetricValue struct {
 	IntValue   int64
-	FloatValue float32
+	FloatValue float64
 	MetricType MetricType
 	ValueType  ValueType
 }
