+++ metrics/sources/kubelet/kubelet.go
@@ -221,7 +221,7 @@ metricloop:
 			mv.IntValue = newest.IntValue
 		case cadvisor.FloatType:
 			mv.ValueType = ValueFloat
-			mv.FloatValue = float32(newest.FloatValue)
+			mv.FloatValue = newest.FloatValue
 		default:
 			glog.V(4).Infof("Skipping %s: unknown custom metric format", spec.Name, spec.Format)
 			continue metricloop
